requirements.txt
torch>=1.12.0
torchvision>=0.13.0
torchmetrics>=0.11.0
matplotlib>=3.3.0
numpy>=1.19.0
tqdm>=4.50.0
pillow>=8.0.0